Implement strings overrides as filters
