+++ src/index.js
@@ -10,7 +10,7 @@ import correctTextFontWeight from './text-font-weight-correct';
 import setupJetpackEditor from './jetpack-editor-setup';
 
 addAction( 'native.pre-render', 'gutenberg-mobile', ( props ) => {
-	require( './ui-strings' );
+	require( './strings-overrides' );
 	correctTextFontWeight();
 	setupJetpackEditor(
 		props.jetpackState || { blogId: 1, isJetpackActive: true }
